Module Vue_js

type 'a vue = 'a Mjs.t

module Make : functor (S : sig ... end) -> sig ... end

Functor to create a vue with typechek of input data

module Router : functor (S : sig ... end) -> sig ... end

Functor to create a vue with typechek of input data and routes

val make : ?⁠computed:('all Mjs.t -> Mjs.any Mjs.optdef) Mjs.table_cons -> ?⁠watch:('all Mjs.t -> Mjs.any -> Mjs.any -> Mjs.any) Mjs.table_cons -> ?⁠methods:Mjs.any Mjs.table_cons -> ?⁠data:'data Mjs.t -> ?⁠lifecycle:(string * ('all Mjs.t -> unit)) list -> ?⁠error_captured:(Mjs.any -> Mjs.any -> string -> bool option) -> ?⁠directives:(string * (string * (Js_of_ocaml.Dom.element Mjs.t -> Vue_component.Internal.binding Mjs.t -> Vue_component.Internal.vnode -> Vue_component.Internal.vnode -> unit)) list) list -> ?⁠filters:(string * (Mjs.any -> Mjs.any)) list -> ?⁠components:(Mjs.top, Mjs.top) Vue_component.Internal.component_arg Mjs.t Mjs.table_cons -> ?⁠delimiters:(string * string) -> ?⁠functional:bool -> ?⁠model:(string option * string option) -> ?⁠inherit_attrs:bool -> ?⁠comments:bool -> ?⁠router:Vue_router.router Mjs.t -> ?⁠store:'state Vuex.store Mjs.t -> string -> 'all vue

generic maker

val set_global : string -> 'a -> unit

set a global $ element of the vue

val get_prop : 'a vue -> string -> 'b

get a $ prop element of the vue

val get_ref : 'a vue -> string -> 'b

get an element of $ref

val get_router : 'a vue -> Vue_router.router Mjs.t

get the router of the vue

val get_store : 'a vue -> 'b Vuex.store Mjs.t

get the store of the vue

val get_route : 'a vue -> ('b, 'c) Vue_router.route Mjs.t

get the current route

val emit0 : 'a vue -> string -> unit

emit events

val emit1 : 'a vue -> string -> 'b -> unit

module SPA : functor (S : sig ... end) -> sig ... end

SPA Functor

module Root : functor (S : sig ... end) -> sig ... end

Root Functor