Module Vue_router

type redirect =

| Rd of string

| Rd_name of string

redirect to path or to named route

type push_obj = {

pu_path : string option;

pu_name : string option;

params : (string * string) list option;

query : (string * string) list option;

}

route argument object for the router methods

val empty_arg : push_obj

empty route object argument

type props =

| PrBool of bool

| PrObj of Mjs.any

| PrFn of push_obj -> Mjs.any

props argument can be a bool, and object or a function

module Internal : sig ... end

Internal module

type router = Internal.router Mjs.t

router class type

type push_arg =

| PStr of string

| PObj of push_obj

route argument can be a string or an object

val push : router -> push_arg -> unit

push to new route

val replace : router -> push_arg -> unit

replace to new route

val go : router -> int -> unit

go to n next route

val before_hook : router -> (push_obj -> push_obj -> (unit -> unit) -> unit) -> unit

register hook before each navigation

val resolve_hook : router -> (push_obj -> push_obj -> (unit -> unit) -> unit) -> unit

register hook before each navigation confirmation

val after_hook : router -> (push_obj -> push_obj -> (unit -> unit) -> unit) -> unit

register hook after each navigation

type ('a, 'b) route_ml = ('a, 'b) Internal.route_ml = {

path : string;

component : ('a, 'b) Vue_component.component_args option;

children : (Mjs.top, Mjs.top) Internal.route Mjs.t list option;

name : string option;

components : ('a, 'b) Vue_component.component_args Mjs.table_cons option;

redirect : redirect option;

alias : string option;

props : props option;

hook : (push_obj -> push_obj -> (unit -> unit) -> unit) option;

}

route object

val empty : string -> ('a, 'b) route_ml

empty route with path

val make : ?⁠mode:string -> (Mjs.top, Mjs.top) route_ml list -> router

make router from route list

module type Tables_S = sig ... end

module Tables : Tables_S